Added a tool for getting training data
The tools shows each image, and you press 'z' for unsharp or
'x' for sharp. It generates a csv with various computed values
as well as the recorded sharp/unsharp value.

Probably should use a 1-5 scale or something, and also put in the focal
length and focal distance of the lense. These would make it easier to use
some kind of regression to determine where the thresholds should sit.